Sigo trabajando y creando el Crud para Alerta y Preferencias compartidas para el login

+++ app/src/main/java/com/example/srodenas/example_with_catalogs/domain/alerts/models/Alert.kt
@@ -3,8 +3,8 @@ package com.example.srodenas.example_with_catalogs.domain.alerts.models
 import java.time.LocalDate
 
 data class Alert(
-    val id: Long = 0,
-    val userId: Long,
+    val id: Int = 0,
+    val userId: Int,
     val textShort: String,
     val message: String,
     val alertDate: LocalDate
